+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### HEAD
 
 * Add documentation.
+* Make the sidebar conditional.
 
 ### 1.0.0